Fix MagickImage(color, uint, uint) constructor call in UITest.cs for Magick.NET v14

+++ src/Controls/tests/TestCases.Shared.Tests/UITest.cs
@@ -669,7 +669,7 @@ namespace Microsoft.Maui.TestCases.Tests
 			var bytes = App.Screenshot();
 
 			// Draw a rounded rectangle with the app window bounds as mask
-			using var surface = new MagickImage(MagickColors.Transparent, width, height);
+			using var surface = new MagickImage(MagickColors.Transparent, (uint)width, (uint)height);
 			new Drawables()
 				.RoundRectangle(0, 0, width, height, cornerRadius, cornerRadius)
 				.FillColor(MagickColors.Black)
